Complete overhaul of value function, cost per duration option and parsing logic
Value funtion now uses sum instead of product for the denominator of the
value function to give more weight to the value, which had a tendency to
be overwhelmed when all three other params where included. Normalization
is also used in a more strict way.

The duration-cost-budget option has been changed into cost-per-duration
which should be equivalent but simpler to understand and now affects
both the value function and the budget, as it should do. This also made
the implementation much cleaner, as the duration gets moved into cost
before calling the value function and as such the value function does
not need to be aware of the flag anymore.

The parsing logic has been also improved considerably adding a lot more
sanity checks and tests to cover them.